bug fix in model

+++ models/simple_models.py
@@ -14,8 +14,8 @@ class ZeroModel(nn.Module):
         # we need to differentiate between batching and no batching
         # without batching we get a 3d tensor (with the 1st dimension being the channels)
         if(len(x.shape) == 3):
-            return torch.zeros_like(x[0], requires_grad=False)
+            return torch.zeros_like(x[0], dtype=torch.float, requires_grad=False)
         # with batching we get a 4d tensor (with the 1st dimension being the number of samples)
         else:
-            return torch.zeros_like(x[:,0,...], requires_grad=False)
+            return torch.zeros_like(x[:,0,...], dtype=torch.float, requires_grad=False)
